Pull taxonomy from eBird API.
Changed to pull taxonomy from the eBird API as well as from a CSV. This
also necessitated updating the tests and changing the parsing code. One
of the changes to the parsing code is to move the scientific short code
to its own spot. Unclear if it's even needed as the rules to generate it
never change.

Also changed the 4-letter-code generation rules slightly to be more in
line with eBird.