Work towards generating bounding tiles for a bbox.
Right now the algorithm works for the happy case, which is the bounding
box is well formed.

If it does cross the anti-meridian, it splits the bounding box into two
and gets tiles for that.

When a poorly formed bounding box is found (no AM crossing, but it
should), nothing presently happens. But there is an in-progress fallback
where the 0,0,0 image is downloaded and a bbox (or bboxes) is determined
from it.

This is important because eBird returns invalid bounding boxes (GBIF
did, but they seem to have fixed that bug).