Fixed invalid bbox finding.
eBird's bounding boxes continue being bad, but some are less bad.
Instead of using +/- 180 degrees for the antimeridian, it's somewhere
between +/- 178 and 180. For anything but absurd zoom levels, this
doesn't matter.

But, if it's only on one edge, it's not crossing the antimeridian, so
this keeps from triggering that case.